Create NotificationExcerptComponent
This can now be covered with unit tests since it's a view component.

Part of #11558
