+++ src/api/app/views/webui/users/notifications/_notifications_list.html.haml
@@ -52,7 +52,7 @@
                       = render NotificationActionDescriptionComponent.new(n)
                   .row.d-none.d-md-block.pl-4
                     .col
-                      %p.mt-3.mb-0= notification.excerpt
+                      = render NotificationExcerptComponent.new(n)
   = paginate notifications, views_prefix: 'webui', window: 2, params: { action: 'index', id: nil }
 
 - content_for :ready_function do
