+++ src/api/spec/components/previews/notification_excerpt_component_preview.rb
@@ -1,0 +1,6 @@
+class NotificationExcerptComponentPreview < ViewComponent::Preview
+  # Preview at http://HOST:PORT/rails/view_components/notification_excerpt_component/preview
+  def preview
+    render(NotificationExcerptComponent.new(Notification.last))
+  end
+end
